get rid of TransactionalSqlClientConnection
and merge its behavior into SqlClientConnection
