+++ hibernate-reactive-core/src/main/java/org/hibernate/reactive/service/initiator/ReactiveConnectionPoolProvider.java
@@ -17,7 +17,5 @@ public interface ReactiveConnectionPoolProvider extends Service {
 	 */
 	ReactiveConnection getConnection();
 
-	ReactiveConnection getTransactionalConnection();
-
 	void close();
 }
